feat: Introduce YarnBuilty and Yarn Invoice enhancements
- Added `save()` validations in `YarnBuiltyController` to nullify non-persistent relationships.
- Updated `YarnBuilty` entity to use `CascadeType.ALL` for `companyYarnOrder` relationship.
- Introduced `/search/yarnInvoiceNo` endpoint in `CompanyYarnOrderController` for fetching orders by invoice number.
- Enhanced `CompanyYarnOrder` and `Orders` entities with proper Jackson annotations for bidirectional relationships.
- Implemented `YarnBuiltyForm` component for saving yarn builty on the frontend.
- Created `yarnBuilty` service for handling save and fetch operations.
- Modified `CompanyYarnOrder` to include `yarnInvoiceNo` field and frontend bindings.
- Updated database connection URL and re-enabled `ddl-auto` property in `application.yml`.

+++ src/test/java/com/example/textile/OrdersSaveTest.java
@@ -310,7 +310,7 @@ public class OrdersSaveTest {
         yarnOrderItem.setHsn("54023300");
         yarnOrderItem.setLotNo("K27092");
         yarnOrderItem.setBoxes(16);
-        yarnOrderItem.setQuantity(549.530);
+        yarnOrderItem.setQuantity(BigDecimal.valueOf(549.530));
         yarnOrderItem.setRate(113.00);
         yarnOrderItem.setYarn(yarn);
         yarnOrderItem.setAmount(BigDecimal.valueOf(62096.89));
